fix(supabase): fix pump_pressure type mismatch in RPC grant/revoke
A prior fix changed p_pump_pressure from double precision to integer
in the function signature but missed updating the matching position
in the revoke/grant statements below it, so they referenced a
signature that didn't match what create or replace function actually
produced ("function ... does not exist" on revoke). Verified the full
25-parameter type list now matches exactly across create/revoke/grant.

+++ supabase/admin-full-station-editor.sql
@@ -160,12 +160,12 @@ $$;
 
 revoke all on function admin_update_station(
   text, text, text, text, text, text, text, text, double precision, double precision, text,
-  text, double precision, double precision, boolean, jsonb, text, text, boolean, text,
+  text, double precision, integer, boolean, jsonb, text, text, boolean, text,
   jsonb, double precision, double precision, integer, text
 ) from anon;
 
 grant execute on function admin_update_station(
   text, text, text, text, text, text, text, text, double precision, double precision, text,
-  text, double precision, double precision, boolean, jsonb, text, text, boolean, text,
+  text, double precision, integer, boolean, jsonb, text, text, boolean, text,
   jsonb, double precision, double precision, integer, text
 ) to authenticated;
